react-redux-ify article page
