+++ src/components/Article/CommentContainer.js
@@ -1,0 +1,39 @@
+import CommentInput from './CommentInput';
+import CommentList from './CommentList';
+import React from 'react';
+
+const CommentContainer = props => {
+  if (props.currentUser) {
+    return (
+      <div className="col-xs-12 col-md-8 offset-md-2">
+        <div>
+          <list-errors errors={props.errors}></list-errors>
+          <CommentInput slug={props.slug} currentUser={props.currentUser} />
+        </div>
+
+        <CommentList
+          comments={props.comments}
+          slug={props.slug}
+          currentUser={props.currentUser} />
+      </div>
+    );
+  } else {
+    return (
+      <div className="col-xs-12 col-md-8 offset-md-2">
+        <p>
+          <Link to="login">Sign in</Link>
+          &nbsp;or&nbsp;
+          <Link to="register">sign up</Link>
+          &nbsp;to add comments on this article.
+        </p>
+
+        <CommentList
+          comments={props.comments}
+          slug={props.slug}
+          currentUser={props.currentUser} />
+      </div>
+    );
+  }
+};
+
+export default CommentContainer;
