+++ src/components/Article/CommentList.js
@@ -1,0 +1,22 @@
+import Comment from './Comment';
+import React from 'react';
+
+const CommentList = props => {
+  return (
+    <div>
+      {
+        props.comments.map(comment => {
+          return (
+            <Comment
+              comment={comment}
+              currentUser={props.currentUser}
+              slug={props.slug}
+              key={comment.id} />
+          );
+        })
+      }
+    </div>
+  );
+};
+
+export default CommentList;
